Project-NASA-Mission-Intelligence-Starter/requirements.txt
chromadb==1.5.7
streamlit==1.56.0
pandas==2.2.2

# LangChain + community integrations
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==1.1.13
langchain-huggingface==0.0.3

# Local LLM + embeddings
sentence-transformers==3.0.1
huggingface-hub==0.23.4
transformers==4.41.2
accelerate==0.31.0
sentencepiece==0.2.0

# RAGAS (only relevancy + faithfulness)
ragas==0.4.3

# Misc
python-dotenv==1.0.1
